chore: move to separate project
