+++ src/AddIns/Uno.WinUI.SpellChecking/AssemblyInfo.cs
@@ -1,0 +1,6 @@
+﻿using global::System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("Uno.UI")]
+[assembly: InternalsVisibleTo("SamplesApp")]
+[assembly: InternalsVisibleTo("SamplesApp.Skia.Generic")]
+[assembly: InternalsVisibleTo("SamplesApp.Skia.WebAssembly.Browser")]
