+++ src/AddIns/Uno.WinUI.SpellChecking/Resources/en_US.aff
