+++ src/Uno.UI/AssemblyInfo.skia.cs
@@ -16,4 +16,5 @@ using Uno.UI.Graphics;
 [assembly: InternalsVisibleTo("Uno.UI.Runtime.Skia.AppleUIKit")]
 [assembly: InternalsVisibleTo("SamplesApp.Skia.Generic")]
 [assembly: InternalsVisibleTo("Uno.UI.RuntimeTests.HRApp.Skia")]
+[assembly: InternalsVisibleTo("Uno.WinUI.SpellChecking")]
