+++ src/Uno.UI/UI/Xaml/Documents/UnicodeText.skia.cs
@@ -5,7 +5,6 @@ using System.Diagnostics;
 using System.Globalization;
 using System.IO;
 using System.Linq;
-using System.Reflection;
 using System.Text;
 using Windows.Foundation;
 using Windows.UI.Text;
@@ -107,7 +106,18 @@ internal readonly partial struct UnicodeText : IParsedText
 	private static readonly Dictionary<int, HashSet<IFontCacheUpdateListener>> _codepointToListeners = new();
 	private static readonly Dictionary<string, HashSet<IFontCacheUpdateListener>> _fontFamilyToListeners = new();
 
-	private static List<WordList>? _wordLists;
+	private static readonly Lazy<ISpellCheckingService?> _spellCheckingService = new(() =>
+	{
+		if (ApiExtensibility.CreateInstance<ISpellCheckingService>(typeof(UnicodeText), out var service))
+		{
+			return service;
+		}
+		else
+		{
+			typeof(UnicodeText).LogError()?.Error($"No implementation of {nameof(ISpellCheckingService)} was found. Spell checking will be disabled.");
+			return null;
+		}
+	});
 
 	private readonly Size _size;
 	private readonly TextAlignment _textAlignment;
@@ -216,9 +226,9 @@ internal readonly partial struct UnicodeText : IParsedText
 		CreateSourceTextFromAndToGlyphMapping(_lines, _textIndexToGlyph);
 
 		_wordBoundaries = GetWords(text);
-		if (isSpellCheckEnabled)
+		if (isSpellCheckEnabled && _spellCheckingService.Value is { } spellCheckingService)
 		{
-			_corrections = GetSpellCheckSuggestions(_wordBoundaries, text);
+			_corrections = spellCheckingService.SpellCheck(_wordBoundaries, text);
 		}
 		else
 		{
@@ -1478,144 +1488,15 @@ internal readonly partial struct UnicodeText : IParsedText
 		}
 	}
 
-	private static List<(int correctionStart, int correctionEnd)?> GetSpellCheckSuggestions(List<int> wordBoundaries, string text)
+	public (int replaceIndexStart, int replaceIndexEnd, List<string> suggestions)? GetSpellCheckSuggestions(int correctionStart, int correctionEnd)
 	{
-		var ret = new List<(int correctionStart, int correctionEnd)?>();
-		var start = 0;
-		foreach (var end in wordBoundaries)
+		if (_spellCheckingService.Value is { } spellCheckingService)
 		{
-			var word = text.Substring(start, end - start);
-			var startTrimmedWord = word.TrimStart();
-			var trimmedWord = startTrimmedWord.TrimEnd();
-
-			_wordLists ??= GetWordLists();
-
-			if (trimmedWord.Length > 0 && !trimmedWord.Any(c => char.IsPunctuation(c) || char.IsNumber(c) || char.IsSeparator(c) || char.IsWhiteSpace(c) || char.IsSymbol(c)))
-			{
-				if (_wordLists.Any(wordList => wordList.Check(trimmedWord)))
-				{
-					ret.Add(null);
-				}
-				else
-				{
-					var correctionStart = word.Length - startTrimmedWord.Length;
-					var correctionEnd = word.Length - startTrimmedWord.Length + trimmedWord.Length;
-					ret.Add((correctionStart, correctionEnd));
-				}
-			}
-			else
-			{
-				ret.Add(null);
-			}
-
-			start = end;
+			return spellCheckingService.GetSpellCheckSuggestions(_text, _wordBoundaries, correctionStart, correctionEnd);
 		}
-
-		return ret;
-	}
-
-	private static List<WordList> GetWordLists()
-	{
-		var wordLists = new List<WordList>();
-		var assembly = Assembly.GetAssembly(typeof(UnicodeText))!;
-		var enAff = assembly.GetManifestResourceNames().First(r => r.Contains("en_US.aff"));
-		var enDic = assembly.GetManifestResourceNames().First(r => r.Contains("en_US.dic"));
-		wordLists.Add(WordList.CreateFromStreams(assembly.GetManifestResourceStream(enDic), assembly.GetManifestResourceStream(enAff)));
-
-		if (FeatureConfiguration.TextBox.CustomSpellCheckDictionaries is { } dictionaries)
+		else
 		{
-			foreach (var (dic, aff) in dictionaries)
-			{
-				if (aff != null)
-				{
-					try
-					{
-						wordLists.Add(WordList.CreateFromStreams(dic, aff));
-					}
-					catch (Exception e)
-					{
-						if (typeof(UnicodeText).Log().IsEnabled(LogLevel.Error))
-						{
-							typeof(UnicodeText).Log().Error($"Failed to load dictionary", e);
-						}
-					}
-				}
-			}
+			return null;
 		}
-
-		return wordLists;
-	}
-
-	public (int replaceIndexStart, int replaceIndexEnd, List<string> suggestions) GetSpellCheckSuggestions(int correctionStart, int correctionEnd)
-	{
-		var wordBoundaries = _wordBoundaries;
-		var text = _text;
-		var index = wordBoundaries.BinarySearch(correctionStart);
-		var i = index >= 0 ? index + 1 : ~index;
-
-		_wordLists ??= GetWordLists();
-
-		if (i < wordBoundaries.Count)
-		{
-			var boundary = wordBoundaries[i];
-			var start = i == 0 ? 0 : wordBoundaries[i - 1];
-
-			if (start <= correctionStart && boundary >= correctionEnd)
-			{
-				var word = text.Substring(start, boundary - start);
-				var startTrimmedWord = word.TrimStart();
-				var trimmedWord = startTrimmedWord.TrimEnd();
-				var suggestions = _wordLists.SelectMany(wordList => wordList.Suggest(trimmedWord)).Distinct().OrderBy(w => LevenshteinDistance(w, trimmedWord)).ToList();
-				return (start + word.Length - startTrimmedWord.Length, start + word.Length - startTrimmedWord.Length + trimmedWord.Length, suggestions);
-			}
-		}
-		return (correctionStart, correctionEnd, new List<string>(0));
-	}
-
-	private static int LevenshteinDistance(string s, string t)
-	{
-		if (string.IsNullOrEmpty(s))
-		{
-			return string.IsNullOrEmpty(t) ? 0 : t.Length;
-		}
-
-		if (string.IsNullOrEmpty(t))
-		{
-			return s.Length;
-		}
-
-		var n = s.Length;
-		var m = t.Length;
-
-		var prev = new int[m + 1];
-		var curr = new int[m + 1];
-
-		// Initialize first row
-		for (var j = 0; j <= m; j++)
-		{
-			prev[j] = j;
-		}
-
-		for (var i = 1; i <= n; i++)
-		{
-			curr[0] = i;
-			for (var j = 1; j <= m; j++)
-			{
-				var cost = s[i - 1] == t[j - 1] ? 0 : 1;
-
-				curr[j] = Math.Min(
-					Math.Min(
-						prev[j] + 1,        // deletion
-						curr[j - 1] + 1     // insertion
-					),
-					prev[j - 1] + cost      // substitution
-				);
-			}
-
-			// Swap rows
-			(prev, curr) = (curr, prev);
-		}
-
-		return prev[m];
 	}
 }
