+++ src/Uno.UI/Uno.UI.Skia.csproj
@@ -53,9 +53,6 @@
 
 	<ItemGroup>
 		<EmbeddedResource Include="Resources\NoiseAsset256x256.png" />
-		<!--	Hunspell dictionary from http://downloads.sourceforge.net/wordlist/hunspell-en_US-2020.12.07.zip	-->
-		<EmbeddedResource Include="Resources\en_US.aff" />
-		<EmbeddedResource Include="Resources\en_US.dic" />
 	</ItemGroup>
 
 	<ItemGroup>
